`ayon_server.addons` implement `RezRepo`
This commit add a new class to the `ayon_server.addons` logic that allows
the server to use Rez (https://github.com/AcademySoftwareFoundation/rez)
as a package manager, including a default `rezbuild.py` for packages that
do not include one, these need to follow the AYON Addon template structure.

THIS BREAKS WITH EXISTING ADDONS present in the `/addons` which are not
rez packages.